style: url_generator 패키지 경로 수정
