--- pennyway-infra/src/main/java/kr/co/pennyway/infra/client/aws/s3/FeedUrlGenerator.java
@@ -1,18 +1,0 @@
-package kr.co.pennyway.infra.client.aws.s3;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import kr.co.pennyway.common.util.UUIDUtil;
-
-public class FeedUrlGenerator implements UrlGenerator {
-	@Override
-	public Map<String, String> generate(String type, String ext, String userId, String chatroomId) {
-		Map<String, String> variablesMap = new HashMap<>();
-		variablesMap.put("uuid", UUIDUtil.generateUUID());
-		variablesMap.put("timestamp", String.valueOf(System.currentTimeMillis()));
-		variablesMap.put("ext", ext);
-		variablesMap.put("feed_id", UUIDUtil.generateUUID());
-		return variablesMap;
-	}
-}
